pyproject.toml
[project]
name = "chat-bot"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "aiohttp>=3.12.14",
    "beautifulsoup4>=4.13.4",
    "channels>=4.2.2",
    "django>=5.2.4",
    "haystack-ai>=2.15.2",
    "nltk>=3.9.1",
    "pymupdf>=1.26.3",
    "python-docx>=1.2.0",
    "sentence-transformers>=5.0.0",
    "torch>=2.7.1",
    "transformers[sentencepiece,torch]>=4.53.3",
    "uvicorn[standard]>=0.35.0",
]
